requirements.txt
PyYaml
kagglehub
torchvision
torch
